% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{confin.bio}
\alias{confin.bio}
\title{Calculates intervals}
\usage{
confin.bio(x, y, se, parm, level)
}
\arguments{
\item{x}{a numerical vector of dataset}

\item{y}{numerical vector of other dataset}

\item{se}{constrast variance/standard error of difference equation}

\item{parm}{estimate W-S value}

\item{level}{the confidence level required}
}
\description{
This function finds upper, lower bounds
}
