% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{SEd}
\alias{SEd}
\title{Calculates contrast variance or standard error of difference equation}
\usage{
SEd(x, c, n)
}
\arguments{
\item{x}{a numerical vector of variances}

\item{c}{a numerical vector of coefficents in difference equation}

\item{n}{a numerical vector of sample sizes}
}
\description{
This function calculates standard error of difference equation or numerator of W-S estimate
}
