% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{WS}
\alias{WS}
\title{Calculates Welch-Satterthwaite degrees of freedom (df) and estimate}
\usage{
WS(x, c, n)
}
\arguments{
\item{x}{a numerical vector of variances}

\item{c}{a numerical vector of coefficents in difference equation}

\item{n}{a numerical vector of sample sizes}
}
\description{
This function finds W-S df and estimate
}
